+++ crates/forge/tests/integration/setup_fork.rs
@@ -258,7 +258,7 @@ fn fork_get_block_info_fails() {
             #[test]
             #[fork(url: "{CHEATNET_RPC_URL}", block_id: BlockId::Number(999999999999))]
             fn fork_get_block_info_fails() {{
-                let block_info = starknet::get_block_info().unbox();
+                starknet::get_block_info().unbox();
             }}
         "#
     )