+++ crates/sncast/src/starknet_commands/account/delete.rs
@@ -14,7 +14,7 @@ use std::io::Write;
 pub struct Delete {
     /// Name of the account to be deleted
     #[clap(short, long)]
-    pub name: Option<String>,
+    pub name: String,
 
     /// If passed with false, Scarb profile won't be removed
     #[clap(long, num_args = 1, default_value = "true")]
@@ -38,19 +38,19 @@ pub fn delete(
     network_name: &str,
     yes: bool,
 ) -> Result<AccountDeleteResponse> {
-    let contents = std::fs::read_to_string(path.clone()).context("Couldn't read accounts file")?;
+    let contents = std::fs::read_to_string(path.clone()).context("Failed to read accounts file")?;
     let items: serde_json::Value = serde_json::from_str(&contents)
         .map_err(|_| anyhow!("Failed to parse accounts file at {path}"))?;
 
     if items[&network_name].is_null() {
-        bail!("No accounts defined for network {}", network_name);
+        bail!("No accounts defined for network = {network_name}");
     }
     if items[&network_name][&name].is_null() {
         bail!("Account with name {name} does not exist")
     }
 
-    let mut items: Map<String, serde_json::Value> =
-        serde_json::from_str(&contents).expect("failed to read file { path }");
+    let mut items: Map<String, serde_json::Value> = serde_json::from_str(&contents)
+        .unwrap_or_else(|_| panic!("Failed to read file at path = {path}"));
 
     // Let's ask confirmation
     if !yes {
@@ -66,9 +66,9 @@ pub fn delete(
     // get to the nested object "nested"
     let nested = items
         .get_mut(network_name)
-        .expect("can't find network")
+        .expect("Failed to find network")
         .as_object_mut()
-        .expect("invalid network format");
+        .expect("Failed to convert network");
 
     // now remove the child from there
     nested.remove(name);