+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `--yes` option to `account delete` command that allows to skip confirmation prompt
 
+#### Changed
+
+- Argument `max-fee` in `account deploy` is now optional
+
 ## [0.13.0] - 2023-12-14
 
 ### Forge